split out test_seen and test_seen_api
